Improvements to speed and expandability
Makes some things more modular by moving the exit into a function, making the inputs come from outside of calculator, and changing the if chains to elif.
Added an except KeyboardInterrupt to the main loop to make it stop cleaner.
Comments were rearranged or removed. Some were unnecessary because the function was explained in the code itself. Some were kept just to make it easier to scroll to where you needed.